chore: migrar configuración de contenidos a Astro v6 y nuevos loaders
